requirements.txt
future==0.18.2
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
